requirements.txt

Django==5.2.0
psycopg2-binary==2.9.9
Pillow==10.2.0
django-import-export==4.0.0
djangorestframework==3.14.0
django-allauth==0.60.0
dj-rest-auth==5.0.2
python-decouple==3.8
